refactor(tests): split if_stmt and for_stmt tests into separate files
Split test cases for IfStmt and ForStmt from test_stmt.py into dedicated
test files to improve code organization and maintainability.
