+++ tests/ut/ir/test_hash_equal.py
@@ -257,81 +257,6 @@ class TestStructuralHash:
         hash_assign = ir.structural_hash(assign)
         assert hash_assign != 0
 
-    def test_if_stmt_same_structure_hash(self):
-        """Test IfStmt nodes with same structure hash."""
-        span = ir.Span.unknown()
-        dtype = DataType.INT64
-        x1 = ir.Var("x", ir.ScalarType(dtype), span)
-        y1 = ir.Var("y", ir.ScalarType(dtype), span)
-        x2 = ir.Var("x", ir.ScalarType(dtype), span)
-        y2 = ir.Var("y", ir.ScalarType(dtype), span)
-        condition1 = ir.Eq(x1, y1, dtype, span)
-        condition2 = ir.Eq(x2, y2, dtype, span)
-        assign1 = ir.AssignStmt(x1, y1, span)
-        assign2 = ir.AssignStmt(x2, y2, span)
-
-        if_stmt1 = ir.IfStmt(condition1, [assign1], [], span)
-        if_stmt2 = ir.IfStmt(condition2, [assign2], [], span)
-
-        hash1 = ir.structural_hash(if_stmt1)
-        hash2 = ir.structural_hash(if_stmt2)
-        # Different variable pointers result in different hashes without auto_mapping
-        assert hash1 != hash2
-
-    def test_if_stmt_different_condition_hash(self):
-        """Test IfStmt nodes with different condition hash."""
-        span = ir.Span.unknown()
-        dtype = DataType.INT64
-        x = ir.Var("x", ir.ScalarType(dtype), span)
-        y = ir.Var("y", ir.ScalarType(dtype), span)
-        z = ir.Var("z", ir.ScalarType(dtype), span)
-        condition1 = ir.Eq(x, y, dtype, span)
-        condition2 = ir.Lt(x, z, dtype, span)
-        assign = ir.AssignStmt(x, y, span)
-
-        if_stmt1 = ir.IfStmt(condition1, [assign], [], span)
-        if_stmt2 = ir.IfStmt(condition2, [assign], [], span)
-
-        hash1 = ir.structural_hash(if_stmt1)
-        hash2 = ir.structural_hash(if_stmt2)
-        assert hash1 != hash2
-
-    def test_if_stmt_different_then_body_hash(self):
-        """Test IfStmt nodes with different then_body hash."""
-        span = ir.Span.unknown()
-        dtype = DataType.INT64
-        x = ir.Var("x", ir.ScalarType(dtype), span)
-        y = ir.Var("y", ir.ScalarType(dtype), span)
-        z = ir.Var("z", ir.ScalarType(dtype), span)
-        condition = ir.Eq(x, y, dtype, span)
-        assign1 = ir.AssignStmt(x, y, span)
-        assign2 = ir.AssignStmt(y, z, span)
-
-        if_stmt1 = ir.IfStmt(condition, [assign1], [], span)
-        if_stmt2 = ir.IfStmt(condition, [assign2], [], span)
-
-        hash1 = ir.structural_hash(if_stmt1)
-        hash2 = ir.structural_hash(if_stmt2)
-        assert hash1 != hash2
-
-    def test_if_stmt_different_else_body_hash(self):
-        """Test IfStmt nodes with different else_body hash."""
-        span = ir.Span.unknown()
-        dtype = DataType.INT64
-        x = ir.Var("x", ir.ScalarType(dtype), span)
-        y = ir.Var("y", ir.ScalarType(dtype), span)
-        z = ir.Var("z", ir.ScalarType(dtype), span)
-        condition = ir.Eq(x, y, dtype, span)
-        assign1 = ir.AssignStmt(x, y, span)
-        assign2 = ir.AssignStmt(y, z, span)
-
-        if_stmt1 = ir.IfStmt(condition, [assign1], [assign1], span)
-        if_stmt2 = ir.IfStmt(condition, [assign1], [assign2], span)
-
-        hash1 = ir.structural_hash(if_stmt1)
-        hash2 = ir.structural_hash(if_stmt2)
-        assert hash1 != hash2
-
     def test_yield_stmt_same_structure_hash(self):
         """Test YieldStmt nodes with same structure hash."""
         span = ir.Span.unknown()
@@ -375,85 +300,6 @@ class TestStructuralHash:
 
         hash1 = ir.structural_hash(yield_stmt1)
         hash2 = ir.structural_hash(yield_stmt2)
-        assert hash1 != hash2
-
-    def test_for_stmt_same_structure_hash(self):
-        """Test ForStmt nodes with same structure hash."""
-        span = ir.Span.unknown()
-        dtype = DataType.INT64
-        i1 = ir.Var("i", ir.ScalarType(dtype), span)
-        start1 = ir.ConstInt(0, dtype, span)
-        stop1 = ir.ConstInt(10, dtype, span)
-        step1 = ir.ConstInt(1, dtype, span)
-        assign1 = ir.AssignStmt(i1, start1, span)
-        for_stmt1 = ir.ForStmt(i1, start1, stop1, step1, [assign1], span)
-
-        i2 = ir.Var("i", ir.ScalarType(dtype), span)
-        start2 = ir.ConstInt(0, dtype, span)
-        stop2 = ir.ConstInt(10, dtype, span)
-        step2 = ir.ConstInt(1, dtype, span)
-        assign2 = ir.AssignStmt(i2, start2, span)
-        for_stmt2 = ir.ForStmt(i2, start2, stop2, step2, [assign2], span)
-
-        hash1 = ir.structural_hash(for_stmt1)
-        hash2 = ir.structural_hash(for_stmt2)
-        assert hash1 == hash2
-
-    def test_for_stmt_different_loop_var_hash(self):
-        """Test ForStmt nodes with different loop vars hash differently."""
-        span = ir.Span.unknown()
-        dtype = DataType.INT64
-        i = ir.Var("i", ir.ScalarType(dtype), span)
-        j = ir.Var("j", ir.ScalarType(dtype), span)
-        start = ir.ConstInt(0, dtype, span)
-        stop = ir.ConstInt(10, dtype, span)
-        step = ir.ConstInt(1, dtype, span)
-        assign = ir.AssignStmt(i, start, span)
-
-        for_stmt1 = ir.ForStmt(i, start, stop, step, [assign], span)
-        for_stmt2 = ir.ForStmt(j, start, stop, step, [assign], span)
-
-        hash1 = ir.structural_hash(for_stmt1)
-        hash2 = ir.structural_hash(for_stmt2)
-        assert hash1 != hash2
-
-    def test_for_stmt_different_range_hash(self):
-        """Test ForStmt nodes with different range values hash differently."""
-        span = ir.Span.unknown()
-        dtype = DataType.INT64
-        i = ir.Var("i", ir.ScalarType(dtype), span)
-        start1 = ir.ConstInt(0, dtype, span)
-        stop1 = ir.ConstInt(10, dtype, span)
-        step1 = ir.ConstInt(1, dtype, span)
-        start2 = ir.ConstInt(0, dtype, span)
-        stop2 = ir.ConstInt(20, dtype, span)
-        step2 = ir.ConstInt(1, dtype, span)
-        assign = ir.AssignStmt(i, start1, span)
-
-        for_stmt1 = ir.ForStmt(i, start1, stop1, step1, [assign], span)
-        for_stmt2 = ir.ForStmt(i, start2, stop2, step2, [assign], span)
-
-        hash1 = ir.structural_hash(for_stmt1)
-        hash2 = ir.structural_hash(for_stmt2)
-        assert hash1 != hash2
-
-    def test_for_stmt_different_body_hash(self):
-        """Test ForStmt nodes with different body statements hash differently."""
-        span = ir.Span.unknown()
-        dtype = DataType.INT64
-        i = ir.Var("i", ir.ScalarType(dtype), span)
-        x = ir.Var("x", ir.ScalarType(dtype), span)
-        start = ir.ConstInt(0, dtype, span)
-        stop = ir.ConstInt(10, dtype, span)
-        step = ir.ConstInt(1, dtype, span)
-        assign1 = ir.AssignStmt(i, start, span)
-        assign2 = ir.AssignStmt(i, x, span)
-
-        for_stmt1 = ir.ForStmt(i, start, stop, step, [assign1], span)
-        for_stmt2 = ir.ForStmt(i, start, stop, step, [assign2], span)
-
-        hash1 = ir.structural_hash(for_stmt1)
-        hash2 = ir.structural_hash(for_stmt2)
         assert hash1 != hash2
 
 
@@ -704,89 +550,6 @@ class TestStructuralEquality:
         # Different types, so not equal
         assert not ir.structural_equal(assign, stmt)
 
-    def test_if_stmt_structural_equal(self):
-        """Test structural equality of IfStmt nodes."""
-        span = ir.Span.unknown()
-        dtype = DataType.INT64
-        x1 = ir.Var("x", ir.ScalarType(dtype), span)
-        y1 = ir.Var("y", ir.ScalarType(dtype), span)
-        x2 = ir.Var("x", ir.ScalarType(dtype), span)
-        y2 = ir.Var("y", ir.ScalarType(dtype), span)
-        condition1 = ir.Eq(x1, y1, dtype, span)
-        condition2 = ir.Eq(x2, y2, dtype, span)
-        assign1 = ir.AssignStmt(x1, y1, span)
-        assign2 = ir.AssignStmt(x2, y2, span)
-
-        if_stmt1 = ir.IfStmt(condition1, [assign1], [], span)
-        if_stmt2 = ir.IfStmt(condition2, [assign2], [], span)
-
-        # Different variable pointers, so not equal without auto_mapping
-        assert not ir.structural_equal(if_stmt1, if_stmt2, enable_auto_mapping=False)
-        # With auto_mapping, they should be equal
-        assert ir.structural_equal(if_stmt1, if_stmt2, enable_auto_mapping=True)
-
-    def test_if_stmt_different_condition_not_equal(self):
-        """Test IfStmt nodes with different condition are not equal."""
-        span = ir.Span.unknown()
-        dtype = DataType.INT64
-        x = ir.Var("x", ir.ScalarType(dtype), span)
-        y = ir.Var("y", ir.ScalarType(dtype), span)
-        z = ir.Var("z", ir.ScalarType(dtype), span)
-        condition1 = ir.Eq(x, y, dtype, span)
-        condition2 = ir.Lt(x, z, dtype, span)
-        assign = ir.AssignStmt(x, y, span)
-
-        if_stmt1 = ir.IfStmt(condition1, [assign], [], span)
-        if_stmt2 = ir.IfStmt(condition2, [assign], [], span)
-
-        assert not ir.structural_equal(if_stmt1, if_stmt2)
-
-    def test_if_stmt_different_then_body_not_equal(self):
-        """Test IfStmt nodes with different then_body are not equal."""
-        span = ir.Span.unknown()
-        dtype = DataType.INT64
-        x = ir.Var("x", ir.ScalarType(dtype), span)
-        y = ir.Var("y", ir.ScalarType(dtype), span)
-        z = ir.Var("z", ir.ScalarType(dtype), span)
-        condition = ir.Eq(x, y, dtype, span)
-        assign1 = ir.AssignStmt(x, y, span)
-        assign2 = ir.AssignStmt(y, z, span)
-
-        if_stmt1 = ir.IfStmt(condition, [assign1], [], span)
-        if_stmt2 = ir.IfStmt(condition, [assign2], [], span)
-
-        assert not ir.structural_equal(if_stmt1, if_stmt2)
-
-    def test_if_stmt_different_else_body_not_equal(self):
-        """Test IfStmt nodes with different else_body are not equal."""
-        span = ir.Span.unknown()
-        dtype = DataType.INT64
-        x = ir.Var("x", ir.ScalarType(dtype), span)
-        y = ir.Var("y", ir.ScalarType(dtype), span)
-        z = ir.Var("z", ir.ScalarType(dtype), span)
-        condition = ir.Eq(x, y, dtype, span)
-        assign1 = ir.AssignStmt(x, y, span)
-        assign2 = ir.AssignStmt(y, z, span)
-
-        if_stmt1 = ir.IfStmt(condition, [assign1], [assign1], span)
-        if_stmt2 = ir.IfStmt(condition, [assign1], [assign2], span)
-
-        assert not ir.structural_equal(if_stmt1, if_stmt2)
-
-    def test_if_stmt_different_from_base_stmt_not_equal(self):
-        """Test IfStmt and base Stmt nodes are not equal."""
-        span = ir.Span.unknown()
-        dtype = DataType.INT64
-        x = ir.Var("x", ir.ScalarType(dtype), span)
-        y = ir.Var("y", ir.ScalarType(dtype), span)
-        condition = ir.Eq(x, y, dtype, span)
-        assign = ir.AssignStmt(x, y, span)
-
-        if_stmt = ir.IfStmt(condition, [assign], [], span)
-        stmt = ir.Stmt(span)
-
-        assert not ir.structural_equal(if_stmt, stmt)
-
     def test_yield_stmt_structural_equal(self):
         """Test structural equality of YieldStmt nodes."""
         span = ir.Span.unknown()
@@ -827,106 +590,6 @@ class TestStructuralEquality:
         yield_stmt2 = ir.YieldStmt([x], span)
 
         assert not ir.structural_equal(yield_stmt1, yield_stmt2)
-
-    def test_for_stmt_structural_equal(self):
-        """Test structural equality of ForStmt nodes."""
-        span = ir.Span.unknown()
-        dtype = DataType.INT64
-        i1 = ir.Var("i", ir.ScalarType(dtype), span)
-        start1 = ir.ConstInt(0, dtype, span)
-        stop1 = ir.ConstInt(10, dtype, span)
-        step1 = ir.ConstInt(1, dtype, span)
-        assign1 = ir.AssignStmt(i1, start1, span)
-        for_stmt1 = ir.ForStmt(i1, start1, stop1, step1, [assign1], span)
-
-        i2 = ir.Var("i", ir.ScalarType(dtype), span)
-        start2 = ir.ConstInt(0, dtype, span)
-        stop2 = ir.ConstInt(10, dtype, span)
-        step2 = ir.ConstInt(1, dtype, span)
-        assign2 = ir.AssignStmt(i2, start2, span)
-        for_stmt2 = ir.ForStmt(i2, start2, stop2, step2, [assign2], span)
-
-        assert ir.structural_equal(for_stmt1, for_stmt2)
-
-    def test_for_stmt_different_loop_var_equal(self):
-        """Test ForStmt nodes with different loop vars are not equal."""
-        span = ir.Span.unknown()
-        dtype = DataType.INT64
-        i = ir.Var("i", ir.ScalarType(dtype), span)
-        j = ir.Var("j", ir.ScalarType(dtype), span)
-        start = ir.ConstInt(0, dtype, span)
-        stop = ir.ConstInt(10, dtype, span)
-        step = ir.ConstInt(1, dtype, span)
-        assign = ir.AssignStmt(i, start, span)
-
-        for_stmt1 = ir.ForStmt(i, start, stop, step, [assign], span)
-        for_stmt2 = ir.ForStmt(j, start, stop, step, [assign], span)
-
-        assert ir.structural_equal(for_stmt1, for_stmt2)
-
-    def test_for_stmt_different_range_not_equal(self):
-        """Test ForStmt nodes with different range values are not equal."""
-        span = ir.Span.unknown()
-        dtype = DataType.INT64
-        i = ir.Var("i", ir.ScalarType(dtype), span)
-        start1 = ir.ConstInt(0, dtype, span)
-        stop1 = ir.ConstInt(10, dtype, span)
-        step1 = ir.ConstInt(1, dtype, span)
-        start2 = ir.ConstInt(0, dtype, span)
-        stop2 = ir.ConstInt(20, dtype, span)
-        step2 = ir.ConstInt(1, dtype, span)
-        assign = ir.AssignStmt(i, start1, span)
-
-        for_stmt1 = ir.ForStmt(i, start1, stop1, step1, [assign], span)
-        for_stmt2 = ir.ForStmt(i, start2, stop2, step2, [assign], span)
-
-        assert not ir.structural_equal(for_stmt1, for_stmt2)
-
-    def test_for_stmt_different_body_not_equal(self):
-        """Test ForStmt nodes with different body statements are not equal."""
-        span = ir.Span.unknown()
-        dtype = DataType.INT64
-        i = ir.Var("i", ir.ScalarType(dtype), span)
-        x = ir.Var("x", ir.ScalarType(dtype), span)
-        start = ir.ConstInt(0, dtype, span)
-        stop = ir.ConstInt(10, dtype, span)
-        step = ir.ConstInt(1, dtype, span)
-        assign1 = ir.AssignStmt(i, start, span)
-        assign2 = ir.AssignStmt(i, x, span)
-
-        for_stmt1 = ir.ForStmt(i, start, stop, step, [assign1], span)
-        for_stmt2 = ir.ForStmt(i, start, stop, step, [assign2], span)
-
-        assert not ir.structural_equal(for_stmt1, for_stmt2)
-
-    def test_for_stmt_empty_vs_non_empty_body_not_equal(self):
-        """Test ForStmt nodes with empty and non-empty body lists are not equal."""
-        span = ir.Span.unknown()
-        dtype = DataType.INT64
-        i = ir.Var("i", ir.ScalarType(dtype), span)
-        start = ir.ConstInt(0, dtype, span)
-        stop = ir.ConstInt(10, dtype, span)
-        step = ir.ConstInt(1, dtype, span)
-        assign = ir.AssignStmt(i, start, span)
-
-        for_stmt1 = ir.ForStmt(i, start, stop, step, [], span)
-        for_stmt2 = ir.ForStmt(i, start, stop, step, [assign], span)
-
-        assert not ir.structural_equal(for_stmt1, for_stmt2)
-
-    def test_for_stmt_different_from_base_stmt_not_equal(self):
-        """Test ForStmt and base Stmt nodes are not equal."""
-        span = ir.Span.unknown()
-        dtype = DataType.INT64
-        i = ir.Var("i", ir.ScalarType(dtype), span)
-        start = ir.ConstInt(0, dtype, span)
-        stop = ir.ConstInt(10, dtype, span)
-        step = ir.ConstInt(1, dtype, span)
-        assign = ir.AssignStmt(i, start, span)
-        for_stmt = ir.ForStmt(i, start, stop, step, [assign], span)
-        base_stmt = ir.Stmt(span)
-
-        assert not ir.structural_equal(for_stmt, base_stmt)
 
     def test_yield_stmt_different_from_base_stmt_not_equal(self):
         """Test YieldStmt and base Stmt nodes are not equal."""
@@ -1449,55 +1112,6 @@ class TestAutoMapping:
         hash_without_auto2 = ir.structural_hash(assign2, enable_auto_mapping=False)
         assert hash_without_auto1 != hash_without_auto2
 
-    def test_auto_mapping_with_if_stmt(self):
-        """Test auto mapping with IfStmt."""
-        # Build: if x == y then x = y else y = x
-        x1 = ir.Var("x", ir.ScalarType(DataType.INT64), ir.Span.unknown())
-        y1 = ir.Var("y", ir.ScalarType(DataType.INT64), ir.Span.unknown())
-        condition1 = ir.Eq(x1, y1, DataType.INT64, ir.Span.unknown())
-        assign1_then = ir.AssignStmt(x1, y1, ir.Span.unknown())
-        assign1_else = ir.AssignStmt(y1, x1, ir.Span.unknown())
-        if_stmt1 = ir.IfStmt(condition1, [assign1_then], [assign1_else], ir.Span.unknown())
-
-        # Build: if a == b then a = b else b = a
-        a = ir.Var("a", ir.ScalarType(DataType.INT64), ir.Span.unknown())
-        b = ir.Var("b", ir.ScalarType(DataType.INT64), ir.Span.unknown())
-        condition2 = ir.Eq(a, b, DataType.INT64, ir.Span.unknown())
-        assign2_then = ir.AssignStmt(a, b, ir.Span.unknown())
-        assign2_else = ir.AssignStmt(b, a, ir.Span.unknown())
-        if_stmt2 = ir.IfStmt(condition2, [assign2_then], [assign2_else], ir.Span.unknown())
-
-        assert ir.structural_equal(if_stmt1, if_stmt2, enable_auto_mapping=True)
-        assert not ir.structural_equal(if_stmt1, if_stmt2, enable_auto_mapping=False)
-
-        hash_with_auto1 = ir.structural_hash(if_stmt1, enable_auto_mapping=True)
-        hash_with_auto2 = ir.structural_hash(if_stmt2, enable_auto_mapping=True)
-        assert hash_with_auto1 == hash_with_auto2
-
-        hash_without_auto1 = ir.structural_hash(if_stmt1, enable_auto_mapping=False)
-        hash_without_auto2 = ir.structural_hash(if_stmt2, enable_auto_mapping=False)
-        assert hash_without_auto1 != hash_without_auto2
-
-    def test_auto_mapping_if_stmt_different_structure(self):
-        """Test auto mapping with IfStmt where structure differs."""
-        # Build: if x == y then x = y
-        x1 = ir.Var("x", ir.ScalarType(DataType.INT64), ir.Span.unknown())
-        y1 = ir.Var("y", ir.ScalarType(DataType.INT64), ir.Span.unknown())
-        condition1 = ir.Eq(x1, y1, DataType.INT64, ir.Span.unknown())
-        assign1 = ir.AssignStmt(x1, y1, ir.Span.unknown())
-        if_stmt1 = ir.IfStmt(condition1, [assign1], [], ir.Span.unknown())
-
-        # Build: if a == b then a = b else b = a
-        a = ir.Var("a", ir.ScalarType(DataType.INT64), ir.Span.unknown())
-        b = ir.Var("b", ir.ScalarType(DataType.INT64), ir.Span.unknown())
-        condition2 = ir.Eq(a, b, DataType.INT64, ir.Span.unknown())
-        assign2_then = ir.AssignStmt(a, b, ir.Span.unknown())
-        assign2_else = ir.AssignStmt(b, a, ir.Span.unknown())
-        if_stmt2 = ir.IfStmt(condition2, [assign2_then], [assign2_else], ir.Span.unknown())
-
-        # Different structure (one has else, one doesn't)
-        assert not ir.structural_equal(if_stmt1, if_stmt2, enable_auto_mapping=True)
-
     def test_auto_mapping_with_yield_stmt(self):
         """Test auto mapping with YieldStmt."""
         # Build: yield x, y
@@ -1534,56 +1148,6 @@ class TestAutoMapping:
 
         # Different lengths should not be equal
         assert not ir.structural_equal(yield_stmt1, yield_stmt2, enable_auto_mapping=True)
-
-    def test_auto_mapping_with_for_stmt(self):
-        """Test auto mapping with ForStmt."""
-        # Build: for i in range(0, 10, 1): i = 0
-        i1 = ir.Var("i", ir.ScalarType(DataType.INT64), ir.Span.unknown())
-        start1 = ir.ConstInt(0, DataType.INT64, ir.Span.unknown())
-        stop1 = ir.ConstInt(10, DataType.INT64, ir.Span.unknown())
-        step1 = ir.ConstInt(1, DataType.INT64, ir.Span.unknown())
-        assign1 = ir.AssignStmt(i1, start1, ir.Span.unknown())
-        for_stmt1 = ir.ForStmt(i1, start1, stop1, step1, [assign1], ir.Span.unknown())
-
-        # Build: for j in range(0, 10, 1): j = 0
-        j = ir.Var("j", ir.ScalarType(DataType.INT64), ir.Span.unknown())
-        start2 = ir.ConstInt(0, DataType.INT64, ir.Span.unknown())
-        stop2 = ir.ConstInt(10, DataType.INT64, ir.Span.unknown())
-        step2 = ir.ConstInt(1, DataType.INT64, ir.Span.unknown())
-        assign2 = ir.AssignStmt(j, start2, ir.Span.unknown())
-        for_stmt2 = ir.ForStmt(j, start2, stop2, step2, [assign2], ir.Span.unknown())
-
-        assert ir.structural_equal(for_stmt1, for_stmt2, enable_auto_mapping=True)
-        assert ir.structural_equal(for_stmt1, for_stmt2, enable_auto_mapping=False)
-
-        hash_with_auto1 = ir.structural_hash(for_stmt1, enable_auto_mapping=True)
-        hash_with_auto2 = ir.structural_hash(for_stmt2, enable_auto_mapping=True)
-        assert hash_with_auto1 == hash_with_auto2
-
-        hash_without_auto1 = ir.structural_hash(for_stmt1, enable_auto_mapping=False)
-        hash_without_auto2 = ir.structural_hash(for_stmt2, enable_auto_mapping=False)
-        assert hash_without_auto1 == hash_without_auto2
-
-    def test_auto_mapping_for_stmt_different_structure(self):
-        """Test auto mapping with ForStmt where structures differ."""
-        # Build: for i in range(0, 10, 1): i = 0
-        i1 = ir.Var("i", ir.ScalarType(DataType.INT64), ir.Span.unknown())
-        start1 = ir.ConstInt(0, DataType.INT64, ir.Span.unknown())
-        stop1 = ir.ConstInt(10, DataType.INT64, ir.Span.unknown())
-        step1 = ir.ConstInt(1, DataType.INT64, ir.Span.unknown())
-        assign1 = ir.AssignStmt(i1, start1, ir.Span.unknown())
-        for_stmt1 = ir.ForStmt(i1, start1, stop1, step1, [assign1], ir.Span.unknown())
-
-        # Build: for j in range(0, 20, 1): j = 0
-        j = ir.Var("j", ir.ScalarType(DataType.INT64), ir.Span.unknown())
-        start2 = ir.ConstInt(0, DataType.INT64, ir.Span.unknown())
-        stop2 = ir.ConstInt(20, DataType.INT64, ir.Span.unknown())
-        step2 = ir.ConstInt(1, DataType.INT64, ir.Span.unknown())
-        assign2 = ir.AssignStmt(j, start2, ir.Span.unknown())
-        for_stmt2 = ir.ForStmt(j, start2, stop2, step2, [assign2], ir.Span.unknown())
-
-        # Different stop values should not be equal
-        assert not ir.structural_equal(for_stmt1, for_stmt2, enable_auto_mapping=True)
 
 
 if __name__ == "__main__":
